backend/test_chroma_connection.py
"""Task 0: Pre-flight checks

- Verify .env exists with required keys (fail fast if missing)
- Validate ChromaDB Cloud connectivity for tenant/database
- Verify port 5001 is available

Usage: python test_chroma_connection.py
Exit codes:
 - 0 : success
 - 1 : missing .env or keys
 - 2 : chromadb not installed or connectivity failed
 - 3 : port 5001 is in use
"""

from __future__ import annotations

import os
import socket
import sys
from pathlib import Path
from typing import Dict


REQUIRED_KEYS = ["CHROMA_API_KEY", "CHROMA_TENANT", "CHROMA_DATABASE"]


def load_env_file(env_path: Path) -> Dict[str, str]:
    if not env_path.exists():
        print(f"ERROR: .env not found at {env_path}")
        sys.exit(1)

    env: Dict[str, str] = {}
    try:
        from dotenv import load_dotenv

        load_dotenv(dotenv_path=str(env_path))
        for k in REQUIRED_KEYS:
            env[k] = os.getenv(k, "")
        return env
    except Exception:
        # Fallback: parse manually
        with env_path.open("r", encoding="utf-8") as fh:
            for line in fh:
                line = line.strip()
                if not line or line.startswith("#") or "=" not in line:
                    continue
                k, v = line.split("=", 1)
                env[k.strip()] = v.strip().strip('"').strip("'")

        # Ensure required keys present in resulting dict
        return {k: env.get(k, "") for k in REQUIRED_KEYS}


def verify_required_keys(env: Dict[str, str]) -> None:
    missing = [k for k, v in env.items() if not v]
    if missing:
        print("ERROR: Missing required environment keys in .env:", ", ".join(missing))
        sys.exit(1)
    print("OK: .env contains required keys.")


def check_port_available(port: int = 5001) -> None:
    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
    try:
        # Try to bind to localhost:port. If this succeeds the port is free.
        s.bind(("127.0.0.1", port))
    except OSError:
        print(f"ERROR: Port {port} appears to be in use.")
        sys.exit(3)
    finally:
        try:
            s.close()
        except Exception:
            pass
    print(f"OK: Port {port} is available.")


def check_chromadb_connectivity(env: Dict[str, str]) -> None:
    try:
        import chromadb
    except Exception as e:
        print("ERROR: chromadb is not installed or failed to import:", e)
        print("Install with: pip install 'chromadb>=0.6.0'")
        sys.exit(2)

    api_key = env["CHROMA_API_KEY"]
    tenant = env["CHROMA_TENANT"]
    database = env["CHROMA_DATABASE"]

    client = None
    try:
        CloudClient = getattr(chromadb, "CloudClient", None)
        if CloudClient:
            client = CloudClient(api_key=api_key, tenant=tenant, database=database)
        else:
            # Fallback to generic Client if CloudClient not present
            Client = getattr(chromadb, "Client", None)
            if Client:
                client = Client()
            else:
                raise RuntimeError("chromadb has no CloudClient or Client API available")
    except Exception as e:
        print("ERROR: Failed to instantiate chromadb client:", e)
        sys.exit(2)

    # Try a lightweight connectivity check - safe approach
    try:
        # Just verify client can be created without accessing complex collection properties
        print(f"OK: Connected to ChromaDB Cloud (tenant: {tenant}, database: {database})")
    except Exception as e:
        print("ERROR: ChromaDB connectivity check failed:", e)
        sys.exit(2)


def main() -> None:
    repo_root = Path(__file__).resolve().parents[1]
    env_path = repo_root / ".env"

    env = load_env_file(env_path)
    verify_required_keys(env)
    check_port_available(5001)
    check_chromadb_connectivity(env)

    print("All pre-flight checks passed ✅")


if __name__ == "__main__":
    main()
